feat(src): create repo_json from locally cloned repositories
